space-intel/requirements.txt
# 数据获取
tushare>=1.3.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
pyarrow>=14.0.0  # parquet 支持

# 配置文件
pyyaml>=6.0

# AI 模型调用
dashscope>=1.14.0  # 阿里百炼 GLM5 API